+++ inc/vector.h
@@ -6,7 +6,7 @@
 /*   By: dchrysov <dchrysov@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/05/07 13:37:03 by junjun            #+#    #+#             */
-/*   Updated: 2025/06/25 15:07:56 by dchrysov         ###   ########.fr       */
+/*   Updated: 2025/07/01 14:59:54 by dchrysov         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -23,9 +23,9 @@
  */
 typedef struct s_vec3
 {
-	double	x;
-	double	y;
-	double	z;
+	float	x;
+	float	y;
+	float	z;
 }			t_vec3;
 
 typedef struct s_color
@@ -36,20 +36,20 @@ typedef struct s_color
 }			t_color;
 
 // vector functions
-t_vec3		new_vector(double x, double y, double z);
+t_vec3		new_vector(float x, float y, float z);
 t_vec3		vec_add(t_vec3 a, t_vec3 b);
 t_vec3		vec_sub(t_vec3 a, t_vec3 b);
-t_vec3		vec_scale(t_vec3 v, double s);
-double		vec_dot(t_vec3 a, t_vec3 b);
+t_vec3		vec_scale(t_vec3 v, float s);
+float		vec_dot(t_vec3 a, t_vec3 b);
 t_vec3		vec_cross(t_vec3 a, t_vec3 b);
-double		vec_length(t_vec3 v);
+float		vec_length(t_vec3 v);
 t_vec3		vec_normal(t_vec3 v);
 t_vec3		vec3_reflect(t_vec3 v, t_vec3 normal);
 
 // color functions
-t_color		color_scale(t_color c, double factor);
+t_color		color_scale(t_color c, float factor);
 t_color		color_add(t_color c1, t_color c2);
-t_color		color_blend(t_color c1, t_color c2, double ratio);
+t_color		color_blend(t_color c1, t_color c2, float ratio);
 t_color		color_mult(t_color c1, t_color c2);
 uint32_t	convert_color(t_color color);
 t_color		clamp_color(t_color c);
