+++ src/main.c
@@ -6,7 +6,7 @@
 /*   By: dchrysov <dchrysov@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/04/08 18:58:58 by junjun            #+#    #+#             */
-/*   Updated: 2025/06/25 14:56:08 by dchrysov         ###   ########.fr       */
+/*   Updated: 2025/07/01 14:59:54 by dchrysov         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,7 +15,7 @@
 static void	viewport_init(t_viewport *vp)
 {
 	vp->fov = 60.0 * M_PI / 180.0;
-	vp->aspect_ratio = (double)WIN_WIDTH / (double)WIN_HEIGHT;
+	vp->aspect_ratio = (float)WIN_WIDTH / (float)WIN_HEIGHT;
 	vp->view_width = 0.0;
 	vp->view_height = 0.0;
 	vp->up = new_vector(0, 0, 0);
